# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Trainer Hub Auth Service API
  version: 1.0.0
  description: Authentication and authorization API for Trainer Hub.
servers:
  - url: http://localhost:8081
paths:
  /api/v1/auth/login:
    post:
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
  /api/v1/auth/refresh:
    post:
      summary: Refresh token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
  /api/v1/auth/validate-token:
    post:
      summary: Validate access token
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token validity result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateTokenResponse'
  /api/v1/auth/me:
    get:
      summary: Current authenticated user
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserResponse'
components:
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    PublicUserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        role:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthTokensResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/PublicUserResponse'
    ValidateTokenResponse:
      type: object
      properties:
        valid:
          type: boolean
        error:
          type: string
          nullable: true